requirements.txt
selenium
pytest-selenium
